strnstr and strrchr

+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: krahnama <krahnama@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/10 21:10:55 by krahnama          #+#    #+#             */
-/*   Updated: 2026/05/30 20:45:44 by krahnama         ###   ########.fr       */
+/*   Updated: 2026/05/31 00:31:16 by krahnama         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	size_t	len;
+	int	len;
 
 	len = ft_strlen(s);
 	while (len >= 0)
